Renaminh g test for CL (test is supposed to be at the begining)
